Thu, Aug 17, 2023, 6:41 PM +02:00

+++ week0/problem_set_1/credit/credit.c
@@ -7,9 +7,9 @@
 
 enum CardState {AMEX, MASTERCARD, VISA, INVALID};
 
-long get_card_number(void)
+unsigned long get_card_number(void)
 {
-    unsigned long long num = 0;
+    unsigned long num = 0;
     while (num <= 0)
     {
         num = get_long("Number: ");
